members_app/requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
boto3==1.36.13
botocore==1.36.13
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==5.0.9 # Higher version breaks django-notifications https://github.com/django-notifications/django-notifications/issues/389
django-appconf==1.0.6
django-hyperview==0.0.2
django-imagekit==5.0.0
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-recurrence==1.11.1
django-unfold==0.43.0
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
django-storages==1.14.4
exponent_server_sdk==2.1.0
gunicorn==23.0.0
ics==0.7.2
idna==3.10
jsonfield==3.1.0
kombu==5.4.2
pilkit==3.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
swapper==1.4.0
TatSu==5.12.2
types-python-dateutil==2.9.0.20241206
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==5.0.9
django-appconf==1.0.6
django-hyperview==0.0.2
django-imagekit==5.0.0
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-recurrence==1.11.1
django-tenants==3.7.0
django-unfold==0.43.0
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
exponent_server_sdk==2.1.0
ics==0.7.2
idna==3.10
jsonfield==3.1.0
kombu==5.4.2
pilkit==3.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
swapper==1.4.0
TatSu==5.12.2
types-python-dateutil==2.9.0.20241206
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
